+++ IntroSkipper/Db/DetectionCacheDbContext.cs
@@ -172,6 +172,4 @@ public class DetectionCacheDbContext : DbContext
             return false;
         }
     }
-
 }
-
